# gh-132493: Lazily determine protocol attributes