--- a/Lib/test/test_typing.py
+++ b/Lib/test/test_typing.py
@@ -3137,6 +3137,7 @@ class ProtocolTests(BaseTestCase):
         class Base(Protocol):
             x: int
 
+        @runtime_checkable
         class Child(Base, Protocol):
             y: str