--- a/Lib/test/test_typing.py
+++ b/Lib/test/test_typing.py
@@ -6338,7 +6338,7 @@ class InternalsTests(BaseTestCase):
             "inspect",
             "re",
             "contextlib",
-            # "annotationlib",  # TODO
+            "annotationlib",
         })