--- a/Misc/NEWS.d/next/Library/2025-04-16-06-50-13.gh-issue-132493.XvnL7t.rst
+++ b/Misc/NEWS.d/next/Library/2025-04-16-06-50-13.gh-issue-132493.XvnL7t.rst
@@ -1,1 +1,0 @@
-Speed up the creation of :class:`typing.Protocol` subclasses.