@AlexWaygood 2025-04-16
If I understand correctly, this will significantly slow down the _first_ `isinstance()` call against a runtime-checkable protocol, but following `isinstance()` calls will be unaffected. I think that's a good tradeoff, but it might be worth giving the folks who closely watch the benchmarks a heads-up. It might do "interesting" things to this benchmark: https://github.com/python/pyperformance/blob/main/pyperformance/data-files/benchmarks/bm_typing_runtime_protocols/run_benchmark.py

Since it's a micro-benchmark anyway, we could possibly even adjust that benchmark so that it does an `isinstance()` call to warm the `__protocol_attrs_cache__` cache before the benchmark starts timing things?