@JelleZijlstra 2025-04-16
I think it might already do some warmup runs? That makes sense for the specializing interpreter anyway. I'll post on Discord.