@AlexWaygood 2025-04-16
It's a shame we can't use `functools.cached_property` for properties on metaclasses :-(