@JelleZijlstra 2025-04-16
Michael confirmed that it does a warmup run by default.