@AlexWaygood 2025-04-16
Ah that sounds like an appealing alternative. Let's keep the new regression test you added as well, though!